requirements.txt
transformers>=4.35.0
peft>=0.5.0
torch>=2.0.0
pandas>=1.5.0
numpy>=1.24.0
pillow>=9.0.0
tqdm>=4.65.0
nltk>=3.8.0
wandb>=0.15.0
accelerate>=0.20.0
bitsandbytes>=0.40.0
python-dotenv>=1.0.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
evaluate>=0.4.0
rouge-score>=0.1.2
bert-score>=0.3.13
sacrebleu>=2.3.0
pyyaml>=6.0
